fix bug: print column titles to output file with user-chosen filename
We check if we can write to the output file  during mediateRequestSetAnalysisWindows by opening the target file in append mode. When the output file doesn't already exist, this creates a new empty file.  Ascii output code checked if the file already existed to decide whether to print column titles or not, so column titles were never printed.

Solution: use file size to check if the output file is new.  If file size is 0, we conclude it is a new file and write the column titles.

git-svn-id: http://svn-ae/ae_doas/qdoas/trunk/Src@975 0551a244-8104-0410-bdc1-badfa85f4f5d