improve OMI measurement date calculation.  Calculation based on TAI-93 leads to increasing differences with correct UTC time, as more leap seconds are inserted (8 second difference as of 2012).
The new method uses the UTC date given in the orbit file, and the number of seconds since the start of the orbit. This way, we get the correct UTC time for the measurement, except for those orbits during which a leap seconds occurs. For those orbits, Qdoas time will be 1 second ahead of actual UTC time after the leap second.  This effects only 3 orbits since Aura launch in 2004.

git-svn-id: http://svn-ae/ae_doas/qdoas/trunk/Src@1060 0551a244-8104-0410-bdc1-badfa85f4f5d