Use current timezone to convert time values back.
There is no need to set the timezone variable to GMT when converting
date/time values to offsets and back, because any time zone offset in
the first conversion is canceled in the second conversion, as long as
both conversions use the same local time zone.  Not having to set the
"TZ" environment variables makes the code a lot simpler.

git-svn-id: http://svn-ae/ae_doas/qdoas/trunk/Src@1405 0551a244-8104-0410-bdc1-badfa85f4f5d