Prevent WARNING error messages from hiding FATAL errors.
When ERROR_SetLast only checks the errorStack for existing errors with
the same errorId, a fatal error message can be "hidden" by a previous
non-fatal error of the same type, which then confuses the user
("Analysis is not working, but I don't get a fatal error?").
Therefore we only hide errors when an error with same errorId and
errorType exists.

git-svn-id: http://svn-ae/ae_doas/qdoas/trunk/Src@1414 0551a244-8104-0410-bdc1-badfa85f4f5d