Detect missing initial wavelength calibration.
When no instrument calibration file or reference spectrum with a
calibration is given, pTabFeno->LambdaRef will only contain (0.0) and it
is impossible to load a cross section in the correct wavelength range.
We return the error "missing initial wavelength calibration".

Previous behaviour was to set LambdaRef to a grid of pixel numbers,
which obscured the problem and led to confusing error messages when
ANALYSE_LoadCross tried to interpret the pixel numbers as wavelengths
and read the cross section file between (1.-7.) and (n_wavel+7.)nm:
("file empty or not large enough").

git-svn-id: http://svn-ae/ae_doas/qdoas/trunk/Src@1429 0551a244-8104-0410-bdc1-badfa85f4f5d