Better handling of XTrackQualityFlags.
After previous change, XTrackQualityFlags were not read from the L1 file
when the config did not use them to filter records.  However, users
could still select the field for output, and this would cause
uninitialized memory to be written to output.  Now we check whether
XTrackQualityFlags are present in the L1 file or not, and write fill
values if the user selects the field for output though it is not present
in the file.

git-svn-id: http://svn-ae/ae_doas/qdoas/trunk/Src@1438 0551a244-8104-0410-bdc1-badfa85f4f5d