Move "missing wavelength calibration" error.
We cannot check for the presence of a wavelength calibration in
ANALYSE_LoadCross, because for instruments with reference spectra
embedded in the L1 files (e.g. GOME, SCIA, GOME-2), the reference not
loaded until after ANALYSE_LoadCross.

To remove the need for a wavelength calibration in ANALYSE_LoadCross, we
just remove the (lambda_min, lambda_max) selection there (this probably
hardly makes a difference for performance).

Instead, we can check if a correct wavelength grid is present during
ANALYSE_XsInterpolation.

git-svn-id: http://svn-ae/ae_doas/qdoas/trunk/Src@1445 0551a244-8104-0410-bdc1-badfa85f4f5d