Update build configuration to handle different versions of libraries for different versions of QDOAS.
Because qwt 6.1.2 is not binary compatible with qwt 6.0.2, we need to keep different versions of libqwt.so.6 for the different versions of QDOAS installed at BIRA. Old versions can keep using 6.0.2 while newer versions can use 6.1.2.  To make this possible, we have a common directory DOAS/Programmes/QDOAS/lib_common, and a per-version directory DOAS/Programmes/QDOAS/lib_$$VERSION$$, where specific libraries for this version can be stored. The order of the -L and -rpath options guarantees that libraries in the version-specific directory are used before libraries in the _common directory.

git-svn-id: http://svn-ae/ae_doas/qdoas/trunk/Src@1487 0551a244-8104-0410-bdc1-badfa85f4f5d